// src/lib/types.ts
export interface Course {
  id: string;
  timestamp: string;
  managerName: string;
  phone: string;
  entity: string; // Courses running under
  title: string;
  duration: string;
  targetAudience: string;
  fee: string;
  studentsPerBatch: string;
  classFormat: string;
  schedule: string;
  teachingMethod: string;
  mentorship: string;
  recordings: string;
  certificate: string;
  contentSummary: string;
  modules: string;
  learningOutcomes: string;
  courseOutcomes: string;
  highlights: string;
}